Make AbstractTestIntegrationSmokeTest extend AbstractTestQueries
Make `AbstractTestIntegrationSmokeTest` extend `AbstractTestQueries` instead
of `AbstractTestDistributedQueries`. `AbstractTestIntegrationSmokeTest`
are meant to be run by read-only connectors and
`AbstractTestDistributedQueries` are meant to be run by connectors
supporting read/write, and `AbstractTestQueries` are applicable to
read-only connectors.
