+++ plugin/trino-memsql/src/test/java/io/trino/plugin/memsql/TestMemSqlDistributedQueries.java
@@ -130,15 +130,6 @@ public class TestMemSqlDistributedQueries
     }
 
     @Override
-    public void testLargeIn(int valuesCount)
-    {
-        // Running this tests on MemSQL results in
-        // "Available disk space is below the value of 'minimal_disk_space' global variable (100 MB). This query cannot be executed"
-        // on followup tests
-        throw new SkipException("Running testLargeIn on MemSQL results in out-of-disk space errors");
-    }
-
-    @Override
     public void testInsertUnicode()
     {
         // MemSQL's utf8 encoding is 3 bytes and truncates strings upon encountering a 4 byte sequence
