+++ testing/trino-testing/src/main/java/io/trino/testing/AbstractTestIntegrationSmokeTest.java
@@ -39,7 +39,7 @@ import static org.testng.Assert.assertTrue;
  * @see AbstractTestDistributedQueries
  */
 public abstract class AbstractTestIntegrationSmokeTest
-        extends AbstractTestQueryFramework
+        extends AbstractTestQueries
 {
     /**
      * Ensure the tests are run with {@link DistributedQueryRunner}. E.g. {@link LocalQueryRunner} takes some
@@ -162,12 +162,6 @@ public abstract class AbstractTestIntegrationSmokeTest
     }
 
     @Test
-    public void testLimit()
-    {
-        assertEquals(computeActual("SELECT * FROM orders LIMIT 10").getRowCount(), 10);
-    }
-
-    @Test
     public void testMultipleRangesPredicate()
     {
         // List columns explicitly. Some connectors do not maintain column ordering.
@@ -256,27 +250,6 @@ public abstract class AbstractTestIntegrationSmokeTest
                         "FROM (SELECT name, regionkey, nationkey, count(*) count FROM nation GROUP BY name, regionkey, nationkey) n " +
                         "JOIN customer c ON c.nationkey = n.nationkey " +
                         "JOIN region r ON n.regionkey = r.regionkey");
-    }
-
-    @Test
-    public void testShowSchemas()
-    {
-        MaterializedResult actualSchemas = computeActual("SHOW SCHEMAS").toTestTypes();
-
-        MaterializedResult.Builder resultBuilder = MaterializedResult.resultBuilder(getQueryRunner().getDefaultSession(), VARCHAR)
-                .row(getQueryRunner().getDefaultSession().getSchema().orElse("tpch"));
-
-        assertContains(actualSchemas, resultBuilder.build());
-    }
-
-    @Test
-    public void testShowTables()
-    {
-        MaterializedResult actualTables = computeActual("SHOW TABLES").toTestTypes();
-        MaterializedResult expectedTables = MaterializedResult.resultBuilder(getQueryRunner().getDefaultSession(), VARCHAR)
-                .row("orders")
-                .build();
-        assertContains(actualTables, expectedTables);
     }
 
     @Test
